+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/IEmpresaService.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface IEmpresaService {
 
-    public Empresa findById(int id);
+    public Empresa findById(long id);
     public List<Empresa> findAll();
     public Empresa createEmpresa(Empresa empresa);
     public Empresa updateEmpresa(int id, Empresa empresa);
-    public void deleteEmpresa(int id);
+    public void deleteEmpresa(long id);
 }
